+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [0.4.1]
 - Added `BLEScan::find_device` ([#55](https::/github.com/taks/esp32-nimble/pull/55))
 - Added `BLEAdvertisedDevice::adv_type`, `BLEAdvertisedDevice::adv_flags`
+- Added `BLEAddress::from_str`
 
 ## [0.4.0] - 2023-12-01
 - Added `BLEAdvertising::min_interval`, `BLEAdvertising::max_interval` ([#51](https::/github.com/taks/esp32-nimble/pull/51))